refactor: reorganize critical fix tests into appropriate test files
- Move IsDST tests from critical_fixes_test.go to datetime_test.go
- Move FluentDuration tests to api_enhancements_test.go with updated expectations
- Move Period.Range context tests to period_test.go
- Delete separate critical_fixes_test.go file
- Maintain test organization by functionality, not by implementation timing
- All tests passing with proper integration into existing test suites

This improves maintainability and follows Go testing best practices.
